Reborn: Fix signed/unsigned and type-casting issues (4)
This change casts size, index, and frame arithmetic to the correct unsigned or Real types to avoid compiler warnings and invalid comparisons across the game client, logic, network, and renderer. It also fixes a few float conversions and cleans up JSON macro definitions in the online stats interface to keep the code build-safe and consistent.

+++ Core/GameEngine/Source/GameClient/VideoPlayer.cpp
@@ -362,7 +362,7 @@ const Video* VideoPlayer::getVideo( AsciiString movieTitle )
 //============================================================================
 const Video* VideoPlayer::getVideo( Int index )
 {
-	if (index < 0 || index >= mVideosAvailableForPlay.size()) {
+	if (index < 0 || (size_t)index >= mVideosAvailableForPlay.size()) {
 		return nullptr;
 	}
